Refactored FunctionSpace to separate file
